fix: fusion component
