+++ web/src/components/LikertScaleQuestion/LikertScaleQuestion.tsx
@@ -2,12 +2,13 @@ import { Flex, Radio, RadioGroup, Square, Stack, Text } from '@chakra-ui/react'
 
 export type LikertScaleQuestionProps = {
   n: 5 | 7 | 11
-  question: string
+  question?: string
   leftHand?: string
   rightHand?: string
   text?: string[]
   onChange: (nextValue: string) => void
   value: string
+  direction: 'row' | 'column'
 }
 
 const LikertScaleQuestion = ({
@@ -18,52 +19,46 @@ const LikertScaleQuestion = ({
   text,
   onChange,
   value,
+  direction,
 }: LikertScaleQuestionProps) => {
   // Array of [1, 2, ..., n]
   // https://stackoverflow.com/questions/3746725/how-to-create-an-array-containing-1-n
   const scale = Array.from({ length: n }, (_, i) => i + 1)
   return (
     <Stack alignItems="center">
+      { question &&
       <Stack direction="row">
         <Square size="20px" bg="grayIcon" data-testid="square" />
         <Text data-testid="question">{question}</Text>
       </Stack>
+      }
       <Flex alignItems="center" gap={2}>
-        {leftHand != undefined && <Text data-testid="left">{leftHand}</Text>}
+        {leftHand && <Text data-testid="left">{leftHand}</Text>}
         <RadioGroup onChange={onChange} value={value}>
-          {text == undefined ? (
-            <Stack direction="row">
+
+          {text ? (
+            <Stack direction={direction}>
               {/* https://www.pluralsight.com/guides/how-to-implement-a-component-%22loop%22-with-react 
             create a n scale radio component*/}
               {scale.map((item, _) => {
                 return (
-                  <Radio data-testid={`radio${item}`} value={item} key={item} />
-                )
-              })}
-            </Stack>
-          ) : text.length == 11 ? (
-            <Stack direction="row">
-              {scale.map((item, _) => {
-                return (
                   <Radio data-testid={`radio${item}`} value={item} key={item}>
                     {text[item - 1]}
-                  </Radio>
-                )
+                  </Radio>                )
               })}
             </Stack>
           ) : (
-            <Stack direction="column">
+            <Stack direction={direction}>
               {scale.map((item, _) => {
                 return (
-                  <Radio data-testid={`radio${item}`} value={item} key={item}>
-                    {text[item - 1]}
-                  </Radio>
+                  <Radio data-testid={`radio${item}`} value={item} key={item}/>
                 )
               })}
             </Stack>
           )}
         </RadioGroup>
-        {rightHand != undefined && <Text data-testid="right">{rightHand}</Text>}
+
+        {rightHand && <Text data-testid="right">{rightHand}</Text>}
       </Flex>
     </Stack>
   )
