+++ web/src/components/LikertScaleQuestionField/LikertScaleQuestionField.test.tsx
@@ -17,6 +17,7 @@ describe('LikertScaleQuestionField', () => {
             leftHand="No"
             rightHand="Yes"
             name=""
+            direction='row'
           />
         </Form>
       )
