--- web/src/components/UEQQuestionField/UEQQuestionField.stories.tsx
@@ -1,25 +1,0 @@
-// When you've added props to your component,
-// pass Storybook's `args` through this story to control it from the addons panel:
-//
-// ```tsx
-// import type { ComponentStory } from '@storybook/react'
-//
-// export const generated: ComponentStory<typeof UeqQuestionField> = (args) => {
-//   return <UeqQuestionField {...args} />
-// }
-// ```
-//
-// See https://storybook.js.org/docs/react/writing-stories/args.
-
-import type { ComponentMeta } from '@storybook/react'
-
-import UeqQuestionField from './UeqQuestionField'
-
-export const generated = () => {
-  return <UeqQuestionField />
-}
-
-export default {
-  title: 'Components/UeqQuestionField',
-  component: UeqQuestionField,
-} as ComponentMeta<typeof UeqQuestionField>
