+++ web/src/pages/GroupBPage/GroupBPage.tsx
@@ -68,6 +68,7 @@ const GroupBPage = () => {
               question="Is this image private?"
               leftHand="No"
               rightHand="Yes"
+              direction='row'
             />
             <OpenEndedQuestionField
               question="Which elements do you consider as public in this image?"
