+++ web/src/pages/NPSPage/NPSPage.tsx
@@ -18,6 +18,7 @@ const NpsPage = () => {
               n={11}
               question="How likely are you to recommend this interface to a friend?"
               text={['0', '1', '2', '3', '4', '5', '6', '7', '8', '9', '10']}
+              direction="row"
             />
             <OpenEndedQuestionField
               question="Tell us a bit more about why you chosee this rating"
